# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Heroes social network is a project created to make life easier for superhero fans.",
        "title": "Heroes Social Network",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "url": "https://www.linkedin.com/in/leandro-alcantara-pro"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/heroes": {
            "get": {
                "description": "Get Hero By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "Get Hero By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hero id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HeroResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "description": "Update hero",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "Update Hero",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hero id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "body hero",
                        "name": "hero",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.HeroRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HeroResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create hero",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "Create Hero",
                "parameters": [
                    {
                        "description": "hero",
                        "name": "hero",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.HeroRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HeroResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete Hero By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "Delete Hero By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hero id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/heroes/abilities": {
            "post": {
                "description": "Add Ability for Hero",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heroes"
                ],
                "summary": "Add Ability for Hero",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ability id",
                        "name": "ability",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "hero id",
                        "name": "hero",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/teams": {
            "get": {
                "description": "Get Team By Name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get Team By Name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "team",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TeamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Update Team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Update Team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "team",
                        "name": "teamId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "description": "team",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TeamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete Team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Delete Team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "team",
                        "name": "teamId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.HeroRequest": {
            "type": "object",
            "required": [
                "civilName",
                "hero",
                "heroName",
                "universe"
            ],
            "properties": {
                "civilName": {
                    "type": "string"
                },
                "hero": {
                    "type": "boolean"
                },
                "heroName": {
                    "type": "string"
                },
                "team": {
                    "type": "string"
                },
                "universe": {
                    "type": "string"
                }
            }
        },
        "dto.HeroResponse": {
            "type": "object",
            "properties": {
                "civilName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "hero": {
                    "type": "boolean"
                },
                "heroName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "team": {
                    "type": "string"
                },
                "universe": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dto.TeamRequest": {
            "type": "object",
            "required": [
                "name",
                "universe"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "universe": {
                    "type": "string"
                }
            }
        },
        "dto.TeamResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "universe": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}